test(query): Validate the shipped example config file instead of a hardcoded copy
Replace the in-process e2e test (which layered a hardcoded YAML override for
ports) with one that loads config-query-interceptor.yaml as-is via the OTel
config provider, resolves it against the components it references, validates the
service graph, and asserts the query_interceptor_example extension parsed its
deny/redact policy from the file. No overrides, no ports, no hardcoded config.
Runtime hook behavior stays covered by extension_test.go and reader_decorator_test.go.
